more script cleanup

+++ c
@@ -6,12 +6,16 @@ cd ~/a5
 make
 
 exec<$fname
-numLines=0
 
 while read line
 do
-    ssh -n cm2chou@$line "nohup pkill rt -u cm2chou &"
-    ssh -n cm2chou@$line "nohup a5/w > & ~/test1 &"
+    hostLength=${#line}
+
+    if [ $hostLength -eq 4 ]
+    then
+        ssh -n cm2chou@$line "nohup pkill rt -u cm2chou &"
+        ssh -n cm2chou@$line "nohup a5/w > & ~/test1 &"
+    fi
 done
 
 ./rt
